refactor: 상수를 enum으로 변경하여 코드 가독성 및 유지보수성 개선
